requirements.txt
torch
transformers
Pillow
opencv-python
numpy